changed the location of archie and delete functions and made them accessible on all three main views. And changed the look of the archive button on edit project form
