+++ app/components/projectAssignment/projectUserLabel.tsx
@@ -1,11 +1,15 @@
 import Image from "next/image";
 import React from "react";
 
-import { UPSERT_ASSIGNMENT } from "../../gqlQueries";
+import { DELETE_ASSIGNMENT, UPSERT_ASSIGNMENT } from "../../gqlQueries";
 import { UserLabelProps } from "@/app/typeInterfaces";
 import { AddPersonInline } from "../addPersonInline";
 import { useMutation } from "@apollo/client";
 import { useProjectsDataContext } from "@/app/contexts/projectsDataContext";
+import EllipsisDropdownMenu from "../ellipsisDropdownMenu";
+import EditAssignmentModal from "../userAssignment/editAssignmentModal";
+import { useModal } from "@/app/contexts/modalContext";
+import { useGeneralDataContext } from "@/app/contexts/generalContext";
 
 export const ProjectUserLabel = ({
 	project,
@@ -13,15 +17,60 @@ export const ProjectUserLabel = ({
 	clickHandler,
 }: UserLabelProps) => {
 	const { refetchProjectList } = useProjectsDataContext()
-
+	const { openModal, closeModal } = useModal();
+	const { viewer } = useGeneralDataContext()
 	const isUserTBD = assignment.assignedUser === null;
+	const { projectList, setProjectList, singleProjectPage, setSingleProjectPage } = useProjectsDataContext();
 	const [upsertAssignment] = useMutation(UPSERT_ASSIGNMENT, {
 		errorPolicy: "all",
 		onCompleted() {
 			refetchProjectList();
 		},
 	});
+	const canAssignmentBeDeleted = !assignment.workWeeks.some(
+		(week) => (week.actualHours ?? 0) > 0
+	  );
+	  const showArchiveButton = viewer?.id === assignment.assignedUser?.id && !canAssignmentBeDeleted
+	  const showDeleteButton = !assignment.assignedUser || viewer?.id === assignment.assignedUser?.id && assignment.canBeDeleted && canAssignmentBeDeleted
 
+	const [deleteAssignment] = useMutation(DELETE_ASSIGNMENT, {
+		errorPolicy: "all",
+		onCompleted({ deleteAssignment }) {
+			const deletedAssignmentId = deleteAssignment.id;
+			const updatedAssignments = singleProjectPage?.assignments?.filter(
+				(assignment) => assignment.id !== deletedAssignmentId
+			);
+			singleProjectPage && setSingleProjectPage({
+				...singleProjectPage,
+				assignments: updatedAssignments,
+			});
+			const projectId = singleProjectPage?.id;
+			if (projectList && projectId) {
+				const updatedProjectList = projectList.map((project) => {
+					if (project.id === projectId) {
+						return {
+							...project,
+							assignments: project?.assignments?.filter(
+								(assignment) => assignment.id !== deletedAssignmentId
+							),
+						};
+					}
+					return project;
+				});
+				setProjectList(updatedProjectList);
+			}
+		},
+	});
+
+	const handleDeleteAssignmentClick = () => {
+		if (assignment.canBeDeleted) {
+			const variables = {
+				assignmentId: assignment.id,
+			};
+			deleteAssignment({ variables });
+			return;
+	};
+	}
 	const onChangeStatusButtonClick = async () => {
 		const variables = {
 			id: assignment.id,
@@ -33,6 +82,58 @@ export const ProjectUserLabel = ({
 			variables
 		})
 	}
+	const handleArchiveAssignmentClick = async () => {
+		const variables = {
+			id: assignment.id,
+			projectId: project?.id,
+			userId: assignment.assignedUser.id,
+			status: 'archived',
+		};
+		await upsertAssignment({
+			variables
+		})
+	}
+	const displayDeleteOrArchive = () => {
+		if (showArchiveButton) {
+			return {
+				component: <button onClick={handleArchiveAssignmentClick} className="text-gray-900 block px-4 py-2 text-sm">Archive</button>,
+				show: true,
+			}
+		}
+		if (showDeleteButton){
+			return {
+				component: <button onClick={handleDeleteAssignmentClick} className="text-gray-900 block px-4 py-2 text-sm">Delete</button>,
+				show: true,
+			}
+		}
+		return {}
+	}
+	const assignmentDropMenuOptions = [
+		{
+			component: (
+				<button
+					className="text-gray-900 block px-4 py-2 text-sm"
+					onClick={() =>
+						openModal(
+							<EditAssignmentModal
+								assignment={assignment}
+								project={project}
+								closeModal={closeModal}
+							/>
+						)
+					}
+				>
+					Edit Assignment
+				</button>
+			),
+			show: true,
+		},
+		{
+			component: displayDeleteOrArchive().component,
+			show: true,
+		}
+	];
+
 	return (
 		<td className='px-0 pr-0 pt-2 pb-2 font-normal flex align-center sm:w-1/3 w-1/2'>
 			<div
@@ -68,6 +169,10 @@ export const ProjectUserLabel = ({
 					Actual
 				</div>
 			</div>
+			<EllipsisDropdownMenu
+				options={assignmentDropMenuOptions}
+				textColor={"text-gray-900"}
+			/>
 		</td >
 	);
 };
