+++ app/components/userAssignment/userLabel.tsx
@@ -1,12 +1,153 @@
 import React from "react";
-import { UserLabelProps } from "../../typeInterfaces";
+import { AssignmentType, UserLabelProps } from "../../typeInterfaces";
 import EllipsisDropdownMenu from "../ellipsisDropdownMenu";
 import { useModal } from "@/app/contexts/modalContext";
 import EditAssignmentModal from "./editAssignmentModal";
+import { useProjectsDataContext } from "@/app/contexts/projectsDataContext";
+import { useGeneralDataContext } from "@/app/contexts/generalContext";
+import { useUserDataContext } from "@/app/contexts/userDataContext";
+import { DELETE_ASSIGNMENT, UPSERT_ASSIGNMENT } from "@/app/gqlQueries";
+import { useMutation } from "@apollo/client";
 
-export const UserLabel = ({ assignment, clickHandler }: UserLabelProps) => {
+export const UserLabel = ({ assignment, selectedUser, clickHandler }: UserLabelProps) => {
 	const { openModal, closeModal } = useModal();
+	const { setSingleUserPage, userList, setUserList } = useUserDataContext()
+	const { showSummaries, viewer } = useGeneralDataContext()
+	const { projectList, setProjectList, singleProjectPage, setSingleProjectPage } = useProjectsDataContext();
+
 	const isAssignmentProposed = assignment.status === "proposed";
+	const canAssignmentBeDeleted = !assignment.workWeeks.some(
+		(week) => (week.actualHours ?? 0) > 0
+	  );
+	const showArchiveButton = viewer?.id === assignment.assignedUser?.id && !canAssignmentBeDeleted
+	const showDeleteButton = !assignment.assignedUser || viewer?.id === assignment.assignedUser?.id && assignment.canBeDeleted && canAssignmentBeDeleted
+	const [upsertAssignment] = useMutation(UPSERT_ASSIGNMENT, {
+		errorPolicy: "all",
+		onCompleted({ upsertAssignment }) {
+			const updatedAssignments = singleProjectPage?.assignments?.map((assignment) =>
+				assignment.id === upsertAssignment.id ? upsertAssignment : assignment
+			);
+			const updatedProjectList = projectList?.map((project) => {
+				if (project.id === singleProjectPage?.id) {
+					return {
+						...project,
+						assignments: project?.assignments?.map((assignment) =>
+							assignment.id === upsertAssignment.id ? upsertAssignment : assignment
+						),
+					};
+				}
+				return project;
+			});
+			const updatedUserList = userList?.map((user) => {
+				if (user.id === upsertAssignment.assignedUser.id) {
+					const updatedUserAssignments = user.assignments.map((assignment) =>
+						assignment.id === upsertAssignment.id ? upsertAssignment : assignment
+					);
+
+					return {
+						...user,
+						assignments: updatedUserAssignments,
+					};
+				}
+				return user;
+			});
+
+			setUserList(updatedUserList);
+			singleProjectPage && setSingleProjectPage({
+				...singleProjectPage,
+				assignments: updatedAssignments,
+			});
+			setProjectList(updatedProjectList);
+		},
+
+	});
+
+
+	const [deleteAssignment] = useMutation(DELETE_ASSIGNMENT, {
+		errorPolicy: "all",
+		onCompleted({ deleteAssignment }) {
+			const deletedAssignmentId = deleteAssignment.id;
+			const updatedAssignments = singleProjectPage?.assignments?.filter(
+				(assignment) => assignment.id !== deletedAssignmentId
+			);
+			singleProjectPage && setSingleProjectPage({
+				...singleProjectPage,
+				assignments: updatedAssignments,
+			});
+			const projectId = singleProjectPage?.id;
+			if (projectList && projectId) {
+				const updatedProjectList = projectList.map((project) => {
+					if (project.id === projectId) {
+						return {
+							...project,
+							assignments: project?.assignments?.filter(
+								(assignment) => assignment.id !== deletedAssignmentId
+							),
+						};
+					}
+					return project;
+				});
+				setProjectList(updatedProjectList);
+			}
+		},
+	});
+	const handleArchiveAssignmentClick = () => {
+		if (assignment.project && assignment.project.isTempProject) {
+			const removedTempAssignment = selectedUser?.assignments.filter((a: AssignmentType) => a.id !== assignment.id);
+			const selectedUserData = {
+				...selectedUser,
+				assignments: removedTempAssignment || [],
+				name: selectedUser?.name || "Default Name",
+				avatarUrl: selectedUser?.avatarUrl || "defaultAvatarUrl.png",
+			};
+			setSingleUserPage(selectedUserData);
+			return;
+		}
+		if (assignment.status !== 'archived') {
+			const projectId = assignment.project.id;
+			const variables = {
+				id: assignment.id,
+				projectId: projectId,
+				userId: assignment.assignedUser.id,
+				status: 'archived',
+			};
+			upsertAssignment({ variables });
+		}
+	};
+
+	const handleDeleteAssignmentClick = () => {
+		if (assignment.canBeDeleted) {
+				const deletedAssignment = selectedUser?.assignments.filter((a: AssignmentType) => a.id !== assignment.id);
+				const selectedUserData = {
+					...selectedUser,
+					assignments: deletedAssignment || [],
+					name: selectedUser?.name || "Default Name",
+					avatarUrl: selectedUser?.avatarUrl || "defaultAvatarUrl.png",
+				};
+				setSingleUserPage(selectedUserData);
+			const variables = {
+				assignmentId: assignment.id,
+			};
+			deleteAssignment({ variables });
+			return;
+		
+	};
+	}
+	const displayDeleteOrArchive = () => {
+		if (showArchiveButton) {
+			return {
+				component: <button onClick={handleArchiveAssignmentClick} className="text-gray-900 block px-4 py-2 text-sm">Archive</button>,
+				show: true,
+			}
+		}
+		if (showDeleteButton){
+			return {
+				component: <button onClick={handleDeleteAssignmentClick} className="text-gray-900 block px-4 py-2 text-sm">Delete</button>,
+				show: true,
+			}
+		}
+		return {}
+	}
 	const assignmentDropMenuOptions = [
 		{
 			component: (
@@ -26,11 +167,10 @@ export const UserLabel = ({ assignment, clickHandler }: UserLabelProps) => {
 			),
 			show: true,
 		},
-		// temporary removal of delete button
-		// {
-		// 	component: () => <button className="text-gray-900 block px-4 py-2 text-sm">Delete</button>,
-		// 	show: true,
-		// },
+		{
+			component: displayDeleteOrArchive().component,
+			show: true,
+		}
 	];
 	return (
 		<div className="sm:ml-auto ml-0 sm:w-24 flex">
