+++ app/components/userSummary.tsx
@@ -1,149 +1,23 @@
 'use client'
 import React from "react";
-import { useMutation } from "@apollo/client";
-import { ArchiveBoxIcon, XCircleIcon } from "@heroicons/react/24/outline";
 
-import { AssignmentType, UserSummaryProps } from "../typeInterfaces";
-import IconButton from "./iconButton";
-import { DELETE_ASSIGNMENT, UPSERT_ASSIGNMENT } from "../gqlQueries";
-import { useUserDataContext } from "../contexts/userDataContext";
-import { useProjectsDataContext } from "../contexts/projectsDataContext";
+import { UserSummaryProps } from "../typeInterfaces";
 import { useGeneralDataContext } from '../contexts/generalContext';
 import { calculatePlan } from "./scrollingCalendar/helpers";
 
 
-const UserSummary: React.FC<UserSummaryProps> = ({ assignment, selectedUser, project }) => {
+const UserSummary: React.FC<UserSummaryProps> = ({ assignment }) => {
 	const burnedHours = assignment.workWeeks.reduce(
 		(acc, curr) => acc + (curr.actualHours ?? 0),
 		0
 	);
-	const { setSingleUserPage, userList, setUserList, singleUserPage } = useUserDataContext()
-	const { showSummaries, viewer } = useGeneralDataContext()
-	const { projectList, setProjectList, singleProjectPage, setSingleProjectPage } = useProjectsDataContext();
+	const { showSummaries } = useGeneralDataContext()
 
-	const [upsertAssignment] = useMutation(UPSERT_ASSIGNMENT, {
-		errorPolicy: "all",
-		onCompleted({ upsertAssignment }) {
-			const updatedAssignments = singleProjectPage?.assignments?.map((assignment) =>
-				assignment.id === upsertAssignment.id ? upsertAssignment : assignment
-			);
-			const updatedProjectList = projectList?.map((project) => {
-				if (project.id === singleProjectPage?.id) {
-					return {
-						...project,
-						assignments: project?.assignments?.map((assignment) =>
-							assignment.id === upsertAssignment.id ? upsertAssignment : assignment
-						),
-					};
-				}
-				return project;
-			});
-			const updatedUserList = userList?.map((user) => {
-				if (user.id === upsertAssignment.assignedUser.id) {
-					const updatedUserAssignments = user.assignments.map((assignment) =>
-						assignment.id === upsertAssignment.id ? upsertAssignment : assignment
-					);
-
-					return {
-						...user,
-						assignments: updatedUserAssignments,
-					};
-				}
-				return user;
-			});
-
-			setUserList(updatedUserList);
-			singleProjectPage && setSingleProjectPage({
-				...singleProjectPage,
-				assignments: updatedAssignments,
-			});
-			setProjectList(updatedProjectList);
-		},
-
-	});
-
-
-	const [deleteAssignment] = useMutation(DELETE_ASSIGNMENT, {
-		errorPolicy: "all",
-		onCompleted({ deleteAssignment }) {
-			const deletedAssignmentId = deleteAssignment.id;
-			const updatedAssignments = singleProjectPage?.assignments?.filter(
-				(assignment) => assignment.id !== deletedAssignmentId
-			);
-			singleProjectPage && setSingleProjectPage({
-				...singleProjectPage,
-				assignments: updatedAssignments,
-			});
-			const projectId = singleProjectPage?.id;
-			if (projectList && projectId) {
-				const updatedProjectList = projectList.map((project) => {
-					if (project.id === projectId) {
-						return {
-							...project,
-							assignments: project?.assignments?.filter(
-								(assignment) => assignment.id !== deletedAssignmentId
-							),
-						};
-					}
-					return project;
-				});
-				setProjectList(updatedProjectList);
-			}
-		},
-	});
-	const handleArchiveItemClick = () => {
-		if (assignment.project && assignment.project.isTempProject) {
-			const removedTempAssignment = selectedUser?.assignments.filter((a: AssignmentType) => a.id !== assignment.id);
-			const selectedUserData = {
-				...selectedUser,
-				assignments: removedTempAssignment || [],
-				name: selectedUser?.name || "Default Name",
-				avatarUrl: selectedUser?.avatarUrl || "defaultAvatarUrl.png",
-			};
-			setSingleUserPage(selectedUserData);
-			return;
-		}
-		if (assignment.status !== 'archived') {
-			const projectId = project ? project.id : assignment.project.id;
-			const variables = {
-				id: assignment.id,
-				projectId: projectId,
-				userId: assignment.assignedUser.id,
-				status: 'archived',
-			};
-			upsertAssignment({ variables });
-		}
-	};
-
-	const handleDeleteAssignmentClick = () => {
-		if (assignment.canBeDeleted) {
-				const deletedAssignment = selectedUser?.assignments.filter((a: AssignmentType) => a.id !== assignment.id);
-				const selectedUserData = {
-					...selectedUser,
-					assignments: deletedAssignment || [],
-					name: selectedUser?.name || "Default Name",
-					avatarUrl: selectedUser?.avatarUrl || "defaultAvatarUrl.png",
-				};
-				setSingleUserPage(selectedUserData);
-			const variables = {
-				assignmentId: assignment.id,
-			};
-			deleteAssignment({ variables });
-			return;
-		
-	};
-	}
 	const summaries = [
 		{ label: 'future plan', value: calculatePlan(assignment, { isFuture: true }), unit: 'hrs' },
 		{ label: 'burned', value: burnedHours, unit: 'hrs', alwaysShow: true },
 		{ label: 'past plan', value: calculatePlan(assignment, { isFuture: false }), unit: 'hrs' },
 	];
-	const canAssignmentBeDeleted = !assignment.workWeeks.some(
-		(week) => (week.actualHours ?? 0) > 0
-	  );
-	const showArchiveButton = viewer?.id === assignment.assignedUser?.id && !canAssignmentBeDeleted
-	const showDeleteButton = !assignment.assignedUser || viewer?.id === assignment.assignedUser?.id && assignment.canBeDeleted && canAssignmentBeDeleted
-
 
 	return (
 		<td className="font-normal py-2 sm:pl-4 pl-0 pr-0 ml-1 sm:ml-0 w-1/2 sm:w-1/6">
@@ -163,22 +37,6 @@ const UserSummary: React.FC<UserSummaryProps> = ({ assignment, selectedUser, pro
 						)}
 					</div>
 				)}
-				{showArchiveButton && <div className="sm:flex hidden items-start justify-center">
-					<IconButton
-						className='text-black flex items-start justify-center text-transparentGrey'
-						onClick={handleArchiveItemClick}
-						Icon={ArchiveBoxIcon}
-						iconSize={'h6 w-6'}
-					/>
-				</div>}
-				{showDeleteButton && <div className="sm:flex hidden items-start justify-center">
-					<IconButton
-						className='text-black flex items-start justify-center text-transparentGrey'
-						onClick={handleDeleteAssignmentClick}
-						Icon={XCircleIcon}
-						iconSize={'h6 w-6'}
-					/>
-				</div>}
 			</div>
 		</td >
 	);
